use an sqlite database to keep track of condor jobs
This commit updates the script to submit condor jobs to now keep track of all
job submissions in an sqlite database. To submit ntuplizer jobs to condor there
is a two step process. First, we run the script create-condor-jobs which
creates the submission files and adds entries to an sqlite database located at
~/state.db. The second step is to actually submit the jobs to condor. To submit
the jobs you can run the submit-condor-jobs script.