add a check that the chi2 is greater than zero
Occasionaly the refitted vertex has a negative chi2 which causes an error when
computing the p-value. Although previously the quality was still returned as
bad since nan > pval_thr evaluted to false, we make sure to print an error so
that in the future it's clear where the error is.