add an offset to the MC impact parameter uncertainty
This commit adds a small offset of +0.0003 to the MC impact parameter
uncertainty in the covariance matrix and then updates all the code to
use this fixed track when doing fits and computing the impact parameter
significance.

Ideally we would also update the off diagonal elements, but that will be
for a future pull request.